fix compile warning

+++ test/cipher_funcs.c
@@ -44,11 +44,11 @@ void encrypt_file(const char *input_filepath, const unsigned char *key,
 	}
 
 	unsigned char buffer[BUFFER_SIZE];
-	int bytes_read    = 0;
+	size_t bytes_read = 0;
 	int bytes_written = 0;
 	while ((bytes_read = fread(buffer, 1, BUFFER_SIZE, input_file)) > 0) {
 		if (EVP_EncryptUpdate(
-				ctx, buffer, &bytes_written, buffer, bytes_read) != 1) {
+				ctx, buffer, &bytes_written, buffer, (int)bytes_read) != 1) {
 			handle_errors();
 		}
 		fwrite(buffer, 1, bytes_written, output_file);
